+++ Pre-GDC-prototype-addiction/Assets/_Scripts/Manager/BuyCardManager.cs
@@ -69,7 +69,7 @@ namespace Manager
         [Header("Move to Scratch-off")] 
         public RectTransform scratchOffButton;
 
-        public static Action<ScratchCardBrand, int, float, Vector3, Sprite> onScratchCardSelected; //Buy Card Action
+        public static Action<ScratchCardBrand, int, int, Vector3, Sprite> onScratchCardSelected; //Buy Card Action
         public static Action onChangeSubmissionStatus; //Give Card Action
         public static Action<float> onSubmitScratchCard;
